Using the same data fetching logic for rest and graphql
